using int64 instead of float

+++ models/account.go
@@ -15,7 +15,7 @@ type Account struct {
 	Counter            null.Int              `json:"counter"`
 	Script             string                `json:"script"`
 	Storage            string                `json:"storage"`
-	Balance            null.Float            `json:"balance"`
+	Balance            null.Int            `json:"balance"`
 	BlockLevel         null.Int              `json:"block_level" sql:"DEFAULT:'-1'::integer"`
 	AccountsCheckpoint []*AccountsCheckpoint `json:"accounts_checkpoint"` // This line is infered from other tables.
 	DelegatedContracts []*DelegatedContract  `json:"delegated_contracts"` // This line is infered from other tables.
